+++ server/src/Communication/Packet.h
@@ -6,11 +6,11 @@ class Packet{
         int type;
         int seqNum;
         int length;
-        float timestamp;
+        int64_t timestamp;
         std::string message;
 
     public:
-        Packet(int type, int seqNum, int length, float timestamp, std::string message);
+        Packet(int type, int seqNum, int length, int64_t timestamp, std::string message);
         Packet(std::string serializedPacket);
         int getType();
         int getSeqNum();
